Documenting Code
* Cleaning up namespaces
* Breaking up large files
* Organizing files by functionality
