+++ com.chartboost.mediation/Runtime/Data/Metrics.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using Newtonsoft.Json;
 
-namespace Chartboost.Placements
+namespace Chartboost
 {
     #nullable enable
     [Serializable]
