+++ com.chartboost.mediation/Runtime/Platforms/ChartboostMediationUnsupported.cs
@@ -1,5 +1,4 @@
 using System.Threading.Tasks;
-using Chartboost.Placements;
 using Chartboost.Requests;
 
 namespace Chartboost.Platforms
