+++ com.chartboost.mediation/Runtime/Requests/ChartboostMediationAdLoadRequest.cs
@@ -1,50 +1,42 @@
 using System.Collections.Generic;
 using Chartboost.AdFormats.Fullscreen;
-using Chartboost.Placements;
+using Chartboost.Events;
+using Chartboost.Utilities;
 
 namespace Chartboost.Requests
 {
+    /// <summary>
+    /// Base class utilized by all load request.
+    /// </summary>
     public class ChartboostMediationAdLoadRequest
     {
+        /// <summary>
+        /// Base constructor for all load requests.
+        /// </summary>
+        /// <param name="placementName"></param>
+        /// <param name="keywords"></param>
         internal ChartboostMediationAdLoadRequest(string placementName, Dictionary<string, string> keywords)
         {
             PlacementName = placementName;
             Keywords = keywords;
         }
 
+        /// <summary>
+        /// The placement name for the ad.
+        /// </summary>
         public string PlacementName { get; }
+        
+        /// <summary>
+        /// The keywords targeted for the ad.
+        /// </summary>
         public Dictionary<string, string> Keywords { get;  }
 
         internal int AssociatedProxy { get; set; }
+        
         ~ChartboostMediationAdLoadRequest()
         {
             // In case that the request never finishes, and no response ever happens, if disposed by GC remove from Cached requests.
             CacheManager.ReleaseFullscreenAdLoadRequest(AssociatedProxy);
         }
-    }
-
-    public class ChartboostMediationFullscreenAdLoadRequest : ChartboostMediationAdLoadRequest
-    {
-        public ChartboostMediationFullscreenAdLoadRequest(string placementName, Dictionary<string, string> keywords) : base(placementName, keywords) { }
-
-        public event ChartboostMediationFullscreenAdEvent DidClick;
-
-        public event ChartboostMediationFullscreenAdEventWithError DidClose;
-        
-        public event ChartboostMediationFullscreenAdEvent DidExpire;
-        
-        public event ChartboostMediationFullscreenAdEvent DidRecordImpression;
-        
-        public event ChartboostMediationFullscreenAdEvent DidReward;
-
-        internal void OnClick(IChartboostMediationFullscreenAd ad) => DidClick?.Invoke(ad);
-
-        internal void OnClose(IChartboostMediationFullscreenAd ad, ChartboostMediationError? error) => DidClose?.Invoke(ad, error);
-
-        internal void OnRecordImpression(IChartboostMediationFullscreenAd ad) => DidRecordImpression?.Invoke(ad);
-
-        internal void OnExpire(IChartboostMediationFullscreenAd ad) => DidExpire?.Invoke(ad);
-
-        internal void OnReward(IChartboostMediationFullscreenAd ad) => DidReward?.Invoke(ad);
     }
 }
